model backend/requirements.txt
fastapi
uvicorn
numpy
Pillow
tensorflow
